Fix unit tests so they pass on Windows (tested on 8.1).

+++ src/test/java/io/vertx/test/core/RedeploymentTest.java
@@ -88,7 +88,13 @@ public class RedeploymentTest extends VertxTestBase {
 
   @Override
   public void tearDown() throws Exception {
-    vertx.fileSystem().deleteRecursiveBlocking(tempDir.getPath(), true);
+    /* One of the temp files is a jar which is added to the classpath to test verticle redeploy on jar
+     * changes.  This jar remains locked by the JVM until all references to this class loader have been
+     * removed and garbage collected.  The following clean up statement fails because either there is
+     * still some floating reference to the class loader or the JVM hasn't yet garbage collected it.
+     */
+    if(!System.getProperty("os.name").startsWith( "Windows" ))
+      vertx.fileSystem().deleteRecursiveBlocking(tempDir.getPath(), true);
   }
 
   @Test
